refactor(Money): change validation codition
- 로또 상금 계산을 위해 0원 이상이면 금액 생성이 가능하도록 변경

+++ src/main/java/lotto/domain/LottoTicketsFactory.java
@@ -2,13 +2,13 @@ package lotto.domain;
 
 import java.util.List;
 import java.util.stream.Collectors;
-import java.util.stream.IntStream;
+import java.util.stream.LongStream;
 
 public class LottoTicketsFactory {
     public static List<LottoTicket> createAuto(Money money) {
         LottoTicketCreatePolicy createPolicy = new AutoCreatePolicy();
 
-        return IntStream.range(0, money.howManyLottoTickets())
+        return LongStream.range(0, money.howManyLottoTickets())
                 .mapToObj(number -> LottoTicket.of(createPolicy))
                 .collect(Collectors.toList());
     }
